Moved json editor stuff to a modal class and started a resolvers file for API handling
