+++ src/plugin-sidebar.js
@@ -25,11 +25,12 @@ import {
 	Icon,
 	FlexItem,
 	PanelBody,
-	Modal,
 } from '@wordpress/components';
 
 import { UpdateThemePanel } from './editor-sidebar/update-panel';
 import { CreateThemePanel } from './editor-sidebar/create-panel';
+import ThemeJsonEditorModal from './editor-sidebar/json-editor-modal';
+
 import {
 	tool,
 	copy,
@@ -38,12 +39,9 @@ import {
 	chevronRight,
 	archive,
 } from '@wordpress/icons';
-import CodeMirror from '@uiw/react-codemirror';
-import { json } from '@codemirror/lang-json';
 
 const CreateBlockThemePlugin = () => {
 	const [ isEditorOpen, setIsEditorOpen ] = useState( false );
-	const [ themeData, setThemeData ] = useState( '' );
 	const { createErrorNotice } = useDispatch( noticesStore );
 
 	const handleSaveClick = () => {
@@ -102,57 +100,6 @@ const CreateBlockThemePlugin = () => {
 		}
 
 		exportTheme();
-	};
-
-	const toggleThemeJsonEditor = async () => {
-		const fetchOptions = {
-			path: '/create-block-theme/v1/get-theme-data',
-			method: 'GET',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-		};
-
-		try {
-			const response = await apiFetch( fetchOptions );
-			const data = JSON.stringify( response, null, 2 );
-			const themeJson = JSON.stringify(
-				JSON.parse( data )?.data,
-				null,
-				2
-			);
-			setThemeData( themeJson );
-		} catch ( e ) {
-			// @todo: handle error
-			setThemeData( '' );
-		}
-
-		setIsEditorOpen( ! isEditorOpen );
-	};
-
-	const Editor = ( { isOpen = true, value, onChange } ) => {
-		if ( ! isOpen ) {
-			return null;
-		}
-
-		return (
-			<Modal
-				title={
-					<>
-						<Icon icon={ 'info' } />{ ' ' }
-						{ __( 'Info', 'create-block-theme' ) }
-					</>
-				}
-				onRequestClose={ toggleThemeJsonEditor }
-			>
-				<CodeMirror
-					extensions={ [ json() ] }
-					value={ value }
-					onChange={ onChange }
-					width="65vw"
-				/>
-			</Modal>
-		);
 	};
 
 	return (
@@ -246,6 +193,22 @@ const CreateBlockThemePlugin = () => {
 										'create-block-theme'
 									) }
 								</Text>
+								<hr></hr>
+								<Button
+									icon={ edit }
+									onClick={ () => setIsEditorOpen( true ) }
+								>
+									{ __(
+										'Edit Theme JSON …',
+										'create-block-theme'
+									) }
+								</Button>
+								<Text variant="muted">
+									{ __(
+										'Open the theme JSON editor to view and edit the theme data. Use with caution.',
+										'create-block-theme'
+									) }
+								</Text>
 							</VStack>
 						</PanelBody>
 					</NavigatorScreen>
@@ -258,19 +221,12 @@ const CreateBlockThemePlugin = () => {
 						<CreateThemePanel />
 					</NavigatorScreen>
 				</NavigatorProvider>
-				<Button
-					icon={ edit }
-					onClick={ toggleThemeJsonEditor }
-					isSecondary
-				>
-					{ __( 'Edit Theme JSON', 'create-block-theme' ) }
-				</Button>
 			</PluginSidebar>
-			<Editor
-				isOpen={ isEditorOpen }
-				value={ themeData }
-				onChange={ ( value ) => setThemeData( value ) }
-			/>
+			{ isEditorOpen && (
+				<ThemeJsonEditorModal
+					onRequestClose={ () => setIsEditorOpen( false ) }
+				/>
+			) }
 		</>
 	);
 };
